requirements.txt
click
Pillow>=9.0
imagehash==4.3.1
